Twenty-three columns that cannot hold a real EVE value; v0.9.57
Follows 0009, which fixed `station_rigs.location_id` — the first of this class
to be caught, and only because a converted module finally ran against Postgres.
Migration 0010 does the rest.

SQLite's INTEGER is a variable-width 64-bit type that treats the declaration as
advisory, so every one of these has stored oversized values correctly for the
life of the project. A Postgres INTEGER is exactly 32 bits and raises
`NumericValueOutOfRange`. Before the migration, all 23 cases failed on Postgres
and none on SQLite.

**The stated criterion did not survive contact, and that is the useful part.**

The plan was to widen whatever ESI declares as int64 — go to the source of
truth rather than guess. Asking the live OpenAPI spec what CCP actually declares
returns int64 for every integer field it has, `type_id` and `group_id` included.
Following that literally means widening the entire schema while feeling
principled about it. A declared type says what a vendor has reserved the right
to send; it says nothing about what they send.

So the ranges got measured instead — static ids from the committed SDE, live
ids sampled from public ESI endpoints:

    type_id                        371,027    0.017% of the ceiling
    solar_system_id             30,030,141    1.4%
    contract_id                234,465,667    10.9%
    corporation_id           2,042,491,468    95.1%
    character_id             2,124,549,094    98.9%
    order_id                 7,407,646,135    3.4x over
    location_id          1,049,982,731,184    489x over
    volume, 7d Tritanium    34,190,149,437    15.9x over

Three things fell out of measuring that reading would not have given.

**The worst offenders were not ids.** `market_price_cache.volume` is seven days
of regional trade and `jita_available` is every Jita sell order's remaining
units summed. Both are over the ceiling today for Tritanium, Pyerite and
Mexallon — which is most of what an industry tool prices — and neither was on
the list, because the list was about ids. Same for the two volume columns in
`station_volume_cache` and `hub_price_cache`.

**`margin_snapshot.item_id` had to be left alone.** It is `margin_watchlist.id`,
our own autoincrement row number, not an EVE asset id. It is the one column
whose name would have swept it in, which is why the widening is targeted by
(table, column) rather than by name.

**`character_id` is the only preventive case, and the test says so.** At 98.9%
a real id still fits, so writing one passes with or without the fix and proves
nothing. The cases write 2,200,000,000 — the id of a character created a little
way ahead, with roughly 23 million signups closing the gap. A test that cannot
fail is not evidence, and dressing a preventive change as a bug fix would have
hidden that.

Left alone with reasons in the migration docstring: `type_id`, `group_id`,
`category_id`, `region_id`, `solar_system_id`, `planet_id`, `contract_id`, every
`id` we mint ourselves, and the count columns. `quantity` is the one to
re-check — ESI declares it int64 and an asset stack can hold billions of units
of a mineral — but nothing was measured over the ceiling, so it stays.

Three mutations, each in the expected shape. Neutering the migration fails **23
Postgres tests and 0 SQLite ones**. Reverting one declaration fails exactly one
test, in `test_migrations.py`, from the guard that keeps declaration and history
in step — the two-artifact lesson from v0.9.56, applied rather than relearned.
And dropping a single column from the migration's list fails exactly one
Postgres case, which is what proves the 23 are independently meaningful rather
than all riding on one blanket effect.

`batch_alter_table` throughout, so it runs on both backends. It changes nothing
in effect on SQLite, but 0009 established that a migration skipping a backend
leaves the declaration and the history disagreeing there. Indexes survive the
SQLite table rebuild — checked, not assumed.

836 tests green, 2 skipped.

+++ app/db/schema.py
@@ -72,7 +72,7 @@ metadata = MetaData()
 
 characters = Table(
     "characters", metadata,
-    Column("character_id", Integer, primary_key=True, autoincrement=False),
+    Column("character_id", BigInteger, primary_key=True, autoincrement=False),
     Column("character_name", Text, nullable=False),
     # Refresh tokens are bound to the client ID that issued them: change
     # EVE_CLIENT_ID and every row here becomes unusable, which is why the
@@ -80,7 +80,7 @@ characters = Table(
     Column("refresh_token", Text, nullable=False),
     Column("access_token", Text),
     Column("token_expires_at", Float),
-    Column("corporation_id", Integer),
+    Column("corporation_id", BigInteger),
     Column("last_sync_at", Float),
     Column("added_at", Float, nullable=False),
 )
@@ -88,7 +88,7 @@ characters = Table(
 app_sessions = Table(
     "app_sessions", metadata,
     Column("session_id", Text, primary_key=True),
-    Column("character_id", Integer, nullable=False),
+    Column("character_id", BigInteger, nullable=False),
     Column("csrf_token", Text, nullable=False),
     Column("created_at", Float, nullable=False),
     Column("last_seen_at", Float, nullable=False),
@@ -98,7 +98,7 @@ app_owner = Table(
     "app_owner", metadata,
     # Single-row table: the first character to log in claims the instance.
     Column("id", Integer, primary_key=True, autoincrement=False),
-    Column("character_id", Integer, nullable=False),
+    Column("character_id", BigInteger, nullable=False),
     Column("claimed_at", Float, nullable=False),
     CheckConstraint("id = 1", name="ck_app_owner_single_row"),
 )
@@ -108,7 +108,7 @@ app_bootstrap = Table(
     # Single-use, ten-minute login links minted by `python -m app.web.bootstrap`
     # — the way back in when the SSO callback registration does not match.
     Column("token", Text, primary_key=True),
-    Column("character_id", Integer, nullable=False),
+    Column("character_id", BigInteger, nullable=False),
     Column("created_at", Float, nullable=False),
 )
 
@@ -130,21 +130,21 @@ app_defaults = Table(
 
 char_assets_cache = Table(
     "char_assets_cache", metadata,
-    Column("character_id", Integer, nullable=False),
+    Column("character_id", BigInteger, nullable=False),
     Column("data_json", Text, nullable=False),
     Column("cached_at", Float),
 )
 
 corp_assets_cache = Table(
     "corp_assets_cache", metadata,
-    Column("corporation_id", Integer, nullable=False),
+    Column("corporation_id", BigInteger, nullable=False),
     Column("data_json", Text, nullable=False),
     Column("cached_at", Float),
 )
 
 char_blueprints_cache = Table(
     "char_blueprints_cache", metadata,
-    Column("character_id", Integer, nullable=False),
+    Column("character_id", BigInteger, nullable=False),
     Column("data_json", Text, nullable=False),
     Column("cached_at", Float),
 )
@@ -160,7 +160,7 @@ char_jobs_cache = Table(
 
 char_skills_cache = Table(
     "char_skills_cache", metadata,
-    Column("character_id", Integer, primary_key=True, autoincrement=False),
+    Column("character_id", BigInteger, primary_key=True, autoincrement=False),
     Column("data_json", Text, nullable=False),
     Column("cached_at", Float, nullable=False),
 )
@@ -172,7 +172,7 @@ char_wallet_cache = Table(
     # any more. Balances stay here rather than moving into the ledger table
     # next to it: two places holding the same number is how they come to
     # disagree, and this one already has a second consumer.
-    Column("character_id", Integer, primary_key=True, autoincrement=False),
+    Column("character_id", BigInteger, primary_key=True, autoincrement=False),
     Column("balance", Float, nullable=False),
     Column("cached_at", Float, nullable=False),
 )
@@ -292,8 +292,8 @@ market_price_cache = Table(
     Column("cached_at", Float),
     # Both arrived as ALTER TABLE ADD COLUMN guarded by a PRAGMA probe. They
     # are ordinary columns of the baseline now.
-    Column("volume", Integer),
-    Column("jita_available", Integer),
+    Column("volume", BigInteger),
+    Column("jita_available", BigInteger),
 )
 
 custom_price_override = Table(
@@ -311,8 +311,8 @@ hub_price_cache = Table(
     Column("type_id", Integer, primary_key=True, autoincrement=False),
     Column("sell_price", Float),
     Column("buy_price", Float),
-    Column("volume", Integer),
-    Column("available", Integer),
+    Column("volume", BigInteger),
+    Column("available", BigInteger),
     Column("cached_at", Float),
 )
 
@@ -327,11 +327,11 @@ price_history_cache = Table(
 
 station_volume_cache = Table(
     "station_volume_cache", metadata,
-    Column("location_id", Integer, primary_key=True, autoincrement=False),
+    Column("location_id", BigInteger, primary_key=True, autoincrement=False),
     Column("type_id", Integer, primary_key=True, autoincrement=False),
-    Column("volume", Integer),
+    Column("volume", BigInteger),
     Column("best_sell", Float),
-    Column("traded_volume", Integer),
+    Column("traded_volume", BigInteger),
     Column("cached_at", Float),
 )
 
@@ -369,7 +369,7 @@ sci_cache = Table(
 
 facility_tax_cache = Table(
     "facility_tax_cache", metadata,
-    Column("facility_id", Integer, primary_key=True, autoincrement=False),
+    Column("facility_id", BigInteger, primary_key=True, autoincrement=False),
     Column("tax_rate", Float, nullable=False),
     Column("cached_at", BigInteger, nullable=False),
 )
@@ -409,7 +409,7 @@ rig_bonuses = Table(
 
 location_name_cache = Table(
     "location_name_cache", metadata,
-    Column("location_id", Integer, primary_key=True, autoincrement=False),
+    Column("location_id", BigInteger, primary_key=True, autoincrement=False),
     Column("name", Text, nullable=False),
     Column("solar_system_id", Integer),
     # `region_id` existed only as an ALTER — it was never in any CREATE TABLE,
@@ -440,7 +440,7 @@ route_jump_cache = Table(
 
 pi_extractor_cache = Table(
     "pi_extractor_cache", metadata,
-    Column("char_id", Integer, primary_key=True, autoincrement=False),
+    Column("char_id", BigInteger, primary_key=True, autoincrement=False),
     Column("planet_id", Integer, primary_key=True, autoincrement=False),
     Column("product_id", Integer, primary_key=True, autoincrement=False),
     Column("char_name", Text),
@@ -562,9 +562,9 @@ public_contracts = Table(
     Column("volume", Float),
     Column("date_expired", Text),
     Column("title", Text),
-    Column("start_location_id", Integer),
-    Column("end_location_id", Integer),
-    Column("issuer_id", Integer),
+    Column("start_location_id", BigInteger),
+    Column("end_location_id", BigInteger),
+    Column("issuer_id", BigInteger),
     Index("idx_pc_region", "region_id"),
 )
